use long for larger parse capacity

+++ 07 Unit testing/StringSum/INaturalNumber.cs
@@ -2,5 +2,5 @@
 
 public interface INaturalNumber
 {
-    int Parse(string input);
+    long Parse(string input);
 }
