+++ 07 Unit testing/StringSum/NaturalNumber.cs
@@ -2,11 +2,11 @@
 
 public class NaturalNumber : INaturalNumber
 {
-    public int Parse(string input)
+    public long Parse(string input)
     {
         if (string.IsNullOrEmpty(input)) return 0;
         if (!input.All(char.IsDigit)) return 0;
 
-        return int.Parse(input);
+        return long.Parse(input);
     }
 }
